+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.1.2.3b3]
+## [Unreleased]## [0.1.2.3b4]
+
+## [0.1.2.3b3] - 2023-07-05
 - dash_app.HostEnvironment is now an IntEnum to allow integer compare
+- dash_common_utils.diff_dashtable_mi .iteritems() replaced by .items() (Pandas 2.0 deprecated)
 
 ## [0.1.2.3b2] - 2023-03-20
 ### Fixed